fix: componentized chosen hotel preview
