+++ public/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Driven.t</title>
+    <link rel="icon" href="./logo.png" />
   </head>
   <body>
     <noscript>Você precisa habilitar JavaScript para poder utilizar esta página.</noscript>
